+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed a bug where users DM'd with an antiraid DM notification would not be called by their proper name.
 - The ability for a deleted message's old contents to be archived is now a premium-only feature.
 - The ability for an updated message's old + new contents to be archived is now a premium-only feature.
+- When using logs from "$log filteredphrases", archival of the user's full message is now a premium-only feature.
 
 ### Version 3.1.1
 - Fixed bug with antiraid where users would not be actioned or logged properly.